requirements.txt
pillow
pytesseract
pdf2image
tqdm
requests